requirements.txt
facenet-pytorch
torch
torchvision
pillow
numpy
opencv-python